# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/v1/branches-report-excel/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting branches report excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Branches Report Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetReportModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/v1/couriers-report-excel/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting couriers report excel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Couriers Report Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "date",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetReportModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/v1/reports/branches": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting branches report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Branches Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start_date",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end_date",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BranchesReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/reports/operators": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting operators report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Operators Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start_date",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end_date",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperatorsReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/reports/shipper": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting shipper report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get Shipper Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start_date",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "end_date",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShipperReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BranchReport": {
            "type": "object",
            "properties": {
                "admin_panel_orders_count": {
                    "type": "integer"
                },
                "app_orders_count": {
                    "type": "integer"
                },
                "bot_orders_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_sum": {
                    "type": "number"
                },
                "total_sum_cash": {
                    "type": "number"
                },
                "total_sum_click": {
                    "type": "number"
                },
                "total_sum_payme": {
                    "type": "number"
                },
                "website_orders_count": {
                    "type": "integer"
                }
            }
        },
        "models.BranchesReport": {
            "type": "object",
            "properties": {
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BranchReport"
                    }
                }
            }
        },
...
        "models.GetReportModel": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                }
            }
        },
...
        "models.OperatorReport": {
            "type": "object",
            "properties": {
                "admin_panel_orders_count": {
                    "type": "integer"
                },
                "app_orders_count": {
                    "type": "integer"
                },
                "average_per_hour": {
                    "type": "number"
                },
                "bot_orders_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_orders_count": {
                    "type": "integer"
                },
                "website_orders_count": {
                    "type": "integer"
                }
            }
        },
        "models.OperatorsReport": {
            "type": "object",
            "properties": {
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OperatorReport"
                    }
                }
            }
        },
...
        "models.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object"
                }
            }
        },
...
        "models.ShipperReport": {
            "type": "object",
            "properties": {
                "admin_panel_orders_count": {
                    "type": "integer"
                },
                "app_orders_count": {
                    "type": "integer"
                },
                "bot_orders_count": {
                    "type": "integer"
                },
                "total_orders_count": {
                    "type": "integer"
                },
                "total_sum": {
                    "type": "number"
                },
                "total_sum_cash": {
                    "type": "number"
                },
                "total_sum_click": {
                    "type": "number"
                },
                "total_sum_payme": {
                    "type": "number"
                },
                "website_orders_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}